prayer/requirements.txt
adal==0.4.5
asn1crypto==0.22.0
azure==2.0.0
azure-batch==3.0.0
azure-common==1.1.6
azure-datalake-store==0.0.13
azure-graphrbac==0.30.0
azure-keyvault==0.3.5
azure-mgmt==1.0.0
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==4.0.0
azure-mgmt-cdn==0.30.3
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==1.0.0
azure-mgmt-containerregistry==0.2.1
azure-mgmt-datalake-analytics==0.1.6
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.6
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.31.0
azure-mgmt-logic==2.1.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==4.1.0
azure-mgmt-resource==1.1.0
azure-mgmt-scheduler==1.1.2
azure-mgmt-sql==0.5.3
azure-mgmt-storage==1.0.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==5.6.130
azure-servicemanagement-legacy==0.20.6
azure-storage==0.20.0
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.7
certifi==2017.4.17
cffi==1.10.0
click==6.7
cryptography==1.9
decorator==4.0.10
diff-match-patch==20121119
Django==1.9
django-allauth==0.22.0
django-appconf==1.0.2
django-avatar==3.1.0
django-braces==1.11.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extensions==1.8.1
django-filter==1.0.1
django-flatpages-tinymce==0.1.1
django-import-export==0.5.1
django-model-utils==2.6.1
django-phonenumber-field==1.3.0
django-realtime==1.1
django-rest-auth==0.9.1
django-social-auth==0.7.25
django-star-ratings==0.5.4
django-storages==1.6.3
django-stored-messages==1.4.0
django-tinymce==2.6.0
django-wkhtmltopdf==3.1.0
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
djoser==0.5.4
docutils==0.13.1
drfdocs==0.0.11
enum34==1.1.6
futures==3.0.5
goslate==1.5.1
html5lib==1.0b8
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
ipython==5.1.0
ipython-genutils==0.1.0
isodate==0.5.4
jmespath==0.9.1
jsonfield==2.0.1
keyring==10.4.0
Markdown==2.6.8
msrest==0.4.11
msrestazure==0.4.11
oauth2==1.9.0.post1
oauthlib==2.0.2
olefile==0.44
pathlib2==2.1.0
pdfkit==0.6.1
pexpect==4.2.1
phonenumbers==8.5.2
phonenumberslite==8.6.0
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.7.1
ptyprocess==0.5.1
pycparser==2.18
PyDictionary==1.5.2
pydot==1.2.3
Pygments==2.1.3
PyJWT==1.4.2
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
tablib==0.11.4
traitlets==4.3.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
xhtml2pdf==0.0.6
